Add game instructions
